Reorganize components into folders
