+++ src/components/Gif/index.js
